Issue #94: created new json serializers
