+++ src/main/java/org/reso/certification/features/payloads/data-availability.feature
@@ -25,13 +25,13 @@ Feature: Payloads Sampling (Web API)
     Given that metadata have been retrieved from the server and validated
     And the metadata contains RESO Standard Resources
     And "payload-samples" has been created in the build directory
-    Then up to 100 records are sampled from each resource with payload samples stored in "payload-samples"
+    Then up to 100000 records are sampled from each resource with payload samples stored in "payload-samples"
 
   @local-resource-sampling @dd-1.7 @payloads-sampling
   Scenario: Non Standard Resource Sampling - Request Data from Each Server Resource
     Given that metadata have been retrieved from the server and validated
     And the metadata contains local resources
-    Then up to 100 records are sampled from each local resource
+    Then up to 100000 records are sampled from each local resource
 
   @payloads-sampling @dd-1.7
   Scenario: A Data Availability Report is Created from Sampled Records
